requirements.test.txt
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0

# HTTP Testing
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Database Testing
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Storage Testing
boto3==1.34.0
botocore==1.34.0
minio==7.2.0

# Security Testing
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Mocking & Fixtures
faker==22.0.0
freezegun==1.4.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.0
isort==5.12.0
bandit==1.7.6

# Type Checking
types-requests==2.31.0.20231231
types-python-jose==3.3.4.8
types-passlib==1.7.7.13
types-freezegun==1.1.10
typing-extensions==4.9.0
email-validator==2.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocstrings==0.24.0

# Monitoring & Telemetry
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-exporter-otlp>=1.23.0
opentelemetry-instrumentation-logging>=0.44b0
opentelemetry-instrumentation>=0.44b0
opentelemetry-semantic-conventions>=0.44b0
opentelemetry-util-http>=0.44b0
opentelemetry-exporter-prometheus>=0.51b0
prometheus-client==0.18.0

# Development
uvicorn==0.27.0
python-dotenv==1.0.0
watchdog==3.0.0
